requirements.txt
py-cord
aiocoingecko
matplotlib
pandas
python-dotenv